Rename types.hpp -> type.hpp
